mailar_engine/requirements.txt
Django==1.9.4
djangorestframework==3.3.3
django-registration-rest-framework==0.3.0
requests==2.9.1
psycopg2==2.6.1
gunicorn==19.4.0
httpretty==0.8.10
